Move featured article logic into own file and use to render home page articles
